[MC-26] [web] chore: move BaseIntervalToggleGroup from indicator-board to indicator directory
