+++ apps/web/app/ui/components/domain/numerical-guidance/indicator/base-interval-toggle-group.tsx
